# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SignalGrid Device Policy & Security API",
    "description": "Enterprise device security, compliance, and policy enforcement platform. Integrates with MDM, NAC, ITSM, and SIEM systems.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@signalgrid.local"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://api.signalgrid.local",
      "description": "Production API",
      "variables": {}
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/api/v1/session/start": {
      "post": {
        "summary": "Start a device session",
        "description": "Initiate a new device session with badge authentication",
        "operationId": "startSession",
        "tags": ["Sessions"],
        "security": [
          {
            "HMACSignature": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BadgeEvent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionDirective"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid signature or timestamp",
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "429": {
            "description": "Too many requests",
            "$ref": "#/components/responses/RateLimitError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from OIDC provider or admin API key"
      },
      "APIKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Admin API key for development (deprecated in production)"
      },
      "HMACSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Signature",
        "description": "HMAC-SHA256 request signature for device endpoints"
      }
    },
    "schemas": {
      "BadgeEvent": {
        "type": "object",
        "required": ["device", "badge", "timestamp"],
        "properties": {
          "device": {
            "type": "object",
            "required": ["deviceId"],
            "properties": {
              "deviceId": {
                "type": "string"
              },
              "deviceSerial": {
                "type": "string"
              },
              "deviceModel": {
                "type": "string"
              }
            }
          },
          "badge": {
            "type": "object",
            "required": ["badgeId"],
            "properties": {
              "badgeId": {
                "type": "string"
              },
              "badgeUid": {
                "type": "string"
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "nonce": {
            "type": "string",
            "description": "Unique value for replay attack prevention"
          }
        }
      },
      "SessionDirective": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "nextAction": {
            "type": "string",
            "enum": ["LAUNCH_APP", "UNLOCK_DEVICE", "WAIT", "ERROR"]
          },
          "bundleId": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "RateLimitError": {
        "description": "Rate limit exceeded",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  }
}